Moving things around and more tidying up.
