+++ Exiled.API/Enums/SpawnLocation.cs
