+++ Exiled.API/Features/Spawn/StaticSpawnPoint.cs
@@ -5,7 +5,7 @@
 // </copyright>
 // -----------------------------------------------------------------------
 
-namespace Exiled.CustomItems.API.Spawn
+namespace Exiled.API.Features.Spawn
 {
     using UnityEngine;
 
@@ -22,5 +22,16 @@ namespace Exiled.CustomItems.API.Spawn
 
         /// <inheritdoc/>
         public override Vector3 Position { get; set; }
+
+        /// <summary>
+        /// Deconstructs the class into usable variables.
+        /// </summary>
+        /// <param name="chance"><inheritdoc cref="Chance"/></param>
+        /// <param name="position"><inheritdoc cref="Position"/></param>
+        public void Deconstruct(out float chance, out Vector3 position)
+        {
+            chance = Chance;
+            position = Position;
+        }
     }
 }
