+++ Exiled.CustomItems/Commands/Info.cs
@@ -12,8 +12,8 @@ namespace Exiled.CustomItems.Commands
 
     using CommandSystem;
 
+    using Exiled.API.Features.Spawn;
     using Exiled.CustomItems.API.Features;
-    using Exiled.CustomItems.API.Spawn;
     using Exiled.Permissions.Extensions;
 
     using NorthwoodLib.Pools;
